Adding C file
Added .C file
First non-working draft of read data function
